PRO-2798 work in progress
