+++ src/main/java/uk/gov/hmcts/probate/health/HealthConfiguration.java
@@ -4,10 +4,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
+import org.springframework.core.io.ClassPathResource;
 import org.springframework.web.client.RestTemplate;
 import uk.gov.hmcts.probate.config.FeeServiceConfiguration;
 import uk.gov.hmcts.probate.config.PDFServiceConfiguration;
-
 
 @Configuration
 public class HealthConfiguration {
@@ -23,12 +24,15 @@ public class HealthConfiguration {
 
     @Value("${idam.service.host}")
     private String idamServiceHost;
-    
+
     @Value("${evidence.management.host}")
     private String evidenceManagementHost;
 
     @Value("${printservice.host}")
     private String printServiceHost;
+
+    @Value("${git.commit.id}")
+    private String commitId;
 
     @Bean
     public SolsHealthIndicator pdfServiceHealthIndicator() {
@@ -55,4 +59,13 @@ public class HealthConfiguration {
         return new SolsHealthIndicator(printServiceHost, restTemplate);
     }
 
+    @Bean
+    public static PropertySourcesPlaceholderConfigurer placeholderConfigurer() {
+        PropertySourcesPlaceholderConfigurer propsConfig
+                = new PropertySourcesPlaceholderConfigurer();
+        propsConfig.setLocation(new ClassPathResource("git.properties"));
+        propsConfig.setIgnoreResourceNotFound(true);
+        propsConfig.setIgnoreUnresolvablePlaceholders(true);
+        return propsConfig;
+    }
 }
