+++ src/main/resources/application.properties
@@ -3,5 +3,7 @@ s2s.enabled=true
 server.error.whitelabel.enabled=false
 
 # Allows gitCommitId to be publically exposed in the health end-point.
+management.security.enabled=false
 management.health.diskspace.enabled=false
 management.health.db.enabled=false
+management.info.git.enabled=false
